+++ actions/create/questions/deploy/keystore_path_existing.js
@@ -1,21 +1,23 @@
-var path = require('path'),
-    fs = require('fs'),
+var pathHelper = require('../../../../lib/helper/path'),
     ks = require('../../../../lib/android/keystore'),
     Configstore = require('configstore'),
     conf = new Configstore('tarifa');
 
 module.exports = {
     dependency: 'android',
-    type:'input',
-    name:'keystore_path',
+    type: 'input',
+    name: 'keystore_path',
     validate: function (answer) {
         var done = this.async();
-        ks.check(path.resolve(answer)).then(function () {
+        ks.check(pathHelper.resolve(answer)).then(function () {
             done(true);
         }, function (reason) {
             done(reason);
         });
     },
-    message:'What is the keystore path?',
-    default:conf.get('keystore_path')
+    filter: function (answer) {
+        return pathHelper.resolve(answer);
+    },
+    message: 'What is the keystore path?',
+    default: conf.get('keystore_path')
 };