+++ actions/create/tasks/fetch-provisioning-file.js
@@ -1,13 +1,14 @@
 var Q = require('q'),
     path = require('path'),
     print = require('../../../lib/helper/print'),
+    pathHelper = require('../../../lib/helper/path'),
     download = require('../../../lib/ios/nomad/provisioning/download'),
     install = require('../../../lib/ios/nomad/provisioning/install');
 
 module.exports = function (r) {
     if (!r.provisioning_profile_name) return Q.resolve(r);
 
-    var downloadDest = path.join(r.path, 'downloaded.mobileprovision');
+    var downloadDest = pathHelper.resolve(r.path, 'downloaded.mobileprovision');
     return download(
         r.apple_id,
         r.apple_developer_team,
@@ -18,11 +19,11 @@ module.exports = function (r) {
     ).then(function () {
         return install(
             downloadDest,
-            true, // remove the downloaded.mobileprovision file after install
+            false, // do not remove the downloaded.mobileprovision file after install
             r.options.verbose
         );
-    }).then(function (profilePath) {
-        r.provisioning_profile_path = profilePath;
+    }).then(function () {
+        r.provisioning_profile_path = downloadDest;
         return r;
     });
 };