+++ actions/create/tasks/tarifa.js
@@ -3,16 +3,17 @@ var Q = require('q'),
     path = require('path'),
     format = require('util').format,
     print = require('../../../lib/helper/print'),
+    pathHelper = require('../../../lib/helper/path'),
     settings = require('../../../lib/settings'),
     pkg = require('../../../package.json'),
     builder = require('../../../lib/builder');
 
 function makeRootDirectory(response) {
     return fs
-      .isDirectory(response.path)
+      .isDirectory(pathHelper.resolve(response.path))
       .then(function(exist) {
         if(exist) return response;
-        return fs.makeDirectory(response.path)
+        return fs.makeDirectory(pathHelper.resolve(response.path))
           .then(function () { return response });
     });
 };
@@ -20,11 +21,11 @@ function makeRootDirectory(response) {
 function copyWWWProject(response) {
     return fs
         // create tarifa web app folder
-        .makeDirectory(path.join(response.path, settings.webAppPath))
+        .makeDirectory(pathHelper.resolve(response.path, settings.webAppPath))
         // copy template project to web app folder
         .then(function () {
             var src = response.www,
-                dest = path.join(response.path, settings.webAppPath);
+                dest = pathHelper.resolve(response.path, settings.webAppPath);
             return fs.copyTree(src, dest);
         })
         .then(function () {
@@ -35,7 +36,7 @@ function copyWWWProject(response) {
 };
 
 function initBuilder(response) {
-    return builder.init(response.path, response.options.verbose).then(function () {
+    return builder.init(pathHelper.resolve(response.path), response.options.verbose).then(function () {
         return response;
     }).fail(function (error) {
         print('Try to run tarifa check when your environment is properly configured.');
@@ -45,7 +46,7 @@ function initBuilder(response) {
 
 function log(response) {
     if (response.options.verbose)
-        print.success('project folders created %s', response.path);
+        print.success('project folders created %s', pathHelper.resolve(response.path));
     return Q.resolve(response);
 };
 
@@ -54,14 +55,14 @@ function createDotTarifaFile(response) {
         current: pkg.version,
         created: pkg.version
     };
-    return fs.write(path.join(response.path, '.tarifa.json'), JSON.stringify(o, null, 2))
+    return fs.write(pathHelper.resolve(response.path, '.tarifa.json'), JSON.stringify(o, null, 2))
         .then(function () { return response; });
 }
 
 function createUserCheckScripts(response) {
     var content = "module.exports = function (msg) {\n    return msg;\n}",
         write = function (platform) {
-            return fs.write(path.join(response.path, 'project/bin', format('check_%s.js', platform)), content);
+            return fs.write(pathHelper.resolve(response.path, 'project/bin', format('check_%s.js', platform)), content);
         };
     return Q.all(response.platforms.map(write))
         .then(function () { return response; });