+++ test/with-devices.js
@@ -1,0 +1,7 @@
+require('./xml/config.xml');
+require('./xml/android/string.xml');
+require('./xml/android/project');
+require('./xml/android/AndroidManifest.xml');
+require('./xml/wp8/csproj');
+require('./xml/wp8/WMAppManifest.xml');
+require('./test')({ run: true });